requirements.txt
pdfplumber
ollama
numpy
faiss-cpu
flashrank